DesktopVK macOS
Draft for a semi-working build of DesktopVK on macOS. Tested on AutoPong and NeonShooter.
Implemented resource fetching ; Resources and libraries need to be bundled into a codeless bundle 'monogame.native.bundle', currently with a placeholder BundleIdentifier. .NET doesn't probe for the bundle or its contents, so it needs to be manually copied to the output directory, and its libraries symlinked there as well so the executable finds them. Need a way to remove the manual effort of copying and symlinking.
Adjustments have been made to accommodate MoltenVK, so unsupported extensions and their features have either been replaced or disabled entirely. Review is still needed for those features.